requirements.txt
--prefer-binary

nicegui==1.4.18
pywebview==5.0.5
send2trash==1.8.2
# auto-py-to-exe
pyinstaller
# pythonnet==2.5.2